requirements.txt
pyTelegramBotAPI==4.11.0
colorama==0.4.6
loguru==0.7.0
python-dotenv==1.0.0
win32-setctime==1.1.0
requests==2.28.2
loader~=2017.9.11
pip==23.1
attrs==23.1.0
response~=0.5.0
Pillow==9.5.0
Jinja2~=3.1.2
keyring==23.13.1
click~=8.1.3
setuptools==67.6.1